UI: Refactor authentication code
