+++ openmetadata-ui/src/main/resources/ui/src/authentication/authenticators/OktaAuthenticator.tsx
@@ -19,9 +19,9 @@ import React, {
   useImperativeHandle,
 } from 'react';
 import { useHistory } from 'react-router-dom';
+import { ROUTES } from '../../constants/constants';
 import { useAuthContext } from '../auth-provider/AuthProvider';
 import { AuthenticatorRef } from '../auth-provider/AuthProvider.interface';
-import { ROUTES } from '../constants/constants';
 
 interface Props {
   children: ReactNode;
