+++ openmetadata-ui/src/main/resources/ui/src/components/dropdown/AnchorDropDownList.tsx
@@ -14,7 +14,7 @@
 import classNames from 'classnames';
 import React, { Fragment } from 'react';
 import { Link } from 'react-router-dom';
-import { useAuthContext } from '../../auth-provider/AuthProvider';
+import { useAuthContext } from '../../authentication/auth-provider/AuthProvider';
 import { useAuth } from '../../hooks/authHooks';
 import SVGIcons from '../../utils/SvgUtils';
 import { DropDownListItem, DropDownListProp } from './types';
