+++ openmetadata-ui/src/main/resources/ui/src/pages/teams/index.tsx
@@ -21,7 +21,7 @@ import { ExtraInfo, FormErrorData } from 'Models';
 import React, { Fragment, useEffect, useState } from 'react';
 import { Link, useHistory, useParams } from 'react-router-dom';
 import AppState from '../../AppState';
-import { useAuthContext } from '../../auth-provider/AuthProvider';
+import { useAuthContext } from '../../authentication/auth-provider/AuthProvider';
 import {
   createTeam,
   deleteTeam,
