+++ openmetadata-ui/src/main/resources/ui/src/router/AppRouter.tsx
@@ -13,7 +13,7 @@
 
 import React from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
-import { useAuthContext } from '../auth-provider/AuthProvider';
+import { useAuthContext } from '../authentication/auth-provider/AuthProvider';
 import Appbar from '../components/app-bar/Appbar';
 import Loader from '../components/Loader/Loader';
 import { ROUTES } from '../constants/constants';
